Updated LED classes.
