+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/LedStates.java
@@ -6,5 +6,6 @@ package org.firstinspires.ftc.teamcode;
 
 public enum LedStates {
     ON,
-    OFF
+    OFF,
+    FLASHING
 }
