+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tests/LedTestOp.java
@@ -1,7 +1,9 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.tests;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+
+import org.firstinspires.ftc.teamcode.submodules.LedController;
 
 /**
  * Created by emper on 3/7/2017.
@@ -9,9 +11,9 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 @Autonomous(name="LedTestOp", group="Comp")
 public class LedTestOp extends LinearOpMode {
-    LedControlFunctionality led;
+    LedController led;
     public void runOpMode() {
-        led = new LedControlFunctionality(this);
+        led = new LedController(this);
         led.LedFlash(30, 0.5);
     }
 }
